+++ src/query.jl
@@ -327,6 +327,279 @@ function count_entities(world::World, q::Query)
 end
 
 """
+    get_components(query::Query, entity::Entity, comp_types::Tuple)
+
+Get the given components for an [Entity](@ref) through a [Query](@ref).
+Components are returned as a tuple.
+
+Only components which are part of the query can be accessed, optional and
+[Const](@ref) ones included. The entity must have all the requested components.
+
+Does not iterate or [close!](@ref close!(::Query)) the query.
+
+# Example
+
+```jldoctest; setup = :(using Ark; include(string(dirname(pathof(Ark)), "/docs.jl"))), output = false
+query = Query(world, (Position, Velocity))
+pos, vel = get_components(query, entity, (Position, Velocity))
+close!(query)
+
+# output
+
+```
+"""
+@inline Base.@constprop :aggressive function get_components(
+    query::Query,
+    entity::Entity,
+    comp_types::Tuple;
+    _unchecked::Bool=false,
+)
+    return @inline _query_get_components(query, entity, _valtuple(comp_types), Val(_unchecked))
+end
+
+"""
+    set_components!(query::Query, entity::Entity, values::Tuple)
+
+Sets the given component values for an [Entity](@ref) through a [Query](@ref).
+Types are inferred from the values.
+
+Only components which are part of the query can be set, optional ones included.
+Components marked as [Const](@ref) in the query are read-only and can't be set.
+The entity must have all the given components.
+
+Does not iterate or [close!](@ref close!(::Query)) the query.
+
+# Example
+
+```jldoctest; setup = :(using Ark; include(string(dirname(pathof(Ark)), "/docs.jl"))), output = false
+query = Query(world, (Position, Velocity))
+set_components!(query, entity, (Position(0, 0), Velocity(1, 1)))
+close!(query)
+
+# output
+
+```
+"""
+@inline Base.@constprop :aggressive function set_components!(
+    query::Query,
+    entity::Entity,
+    values::Tuple;
+    _unchecked::Bool=false,
+)
+    return @inline _query_set_components!(
+        query,
+        entity,
+        Val{typeof(values)}(),
+        values,
+        Val(_unchecked),
+    )
+end
+
+"""
+    has_components(query::Query, entity::Entity, comp_types::Tuple)::Bool
+
+Returns whether an [Entity](@ref) has all the given components.
+
+Only components which are part of the query can be checked, optional and
+[Const](@ref) ones included.
+
+Does not iterate or [close!](@ref close!(::Query)) the query.
+
+# Example
+
+```jldoctest; setup = :(using Ark; include(string(dirname(pathof(Ark)), "/docs.jl"))), output = false
+query = Query(world, (Position, Velocity))
+has = has_components(query, entity, (Position, Velocity))
+close!(query)
+
+# output
+
+```
+"""
+@inline Base.@constprop :aggressive function has_components(
+    query::Query,
+    entity::Entity,
+    comp_types::Tuple;
+    _unchecked::Bool=false,
+)
+    return @inline _query_has_components(query, entity, _valtuple(comp_types), Val(_unchecked))
+end
+
+_query_component_types(::Type{QS}) where {QS<:Tuple} =
+    DataType[_component_type(S) for S in fieldtypes(QS)]
+
+function _query_component_index(query_types::Vector{DataType}, T::DataType)
+    index = findfirst(==(T), query_types)
+    if index === nothing
+        available = join(map(_format_type, query_types), ", ")
+        throw(ArgumentError(lazy"component $(_format_type(T)) is not part of the query on ($available)"))
+    end
+    return index
+end
+
+@noinline function _throw_missing_query_component(::Type{T}) where {T}
+    throw(ArgumentError(lazy"entity has no $T component"))
+end
+
+@inline function _has_query_component(cols::Vector{A}, idx::_EntityIndex) where {A<:AbstractArray}
+    table = Int(idx.table)
+    return table <= length(cols) && length(@inbounds cols[table]) >= idx.row
+end
+
+@inline function _check_query_component(
+    cols::Vector{A},
+    idx::_EntityIndex,
+    ::Type{T},
+) where {A<:AbstractArray,T}
+    if !_has_query_component(cols, idx)
+        _throw_missing_query_component(T)
+    end
+    return nothing
+end
+
+@generated function _query_get_components(
+    q::Query{QS},
+    entity::Entity,
+    ::TS,
+    ::Val{Unchecked},
+) where {QS<:Tuple,TS<:Tuple,Unchecked}
+    types = _to_types(TS)
+    _check_no_duplicates(types)
+
+    if length(types) == 0
+        return :(())
+    end
+
+    query_types = _query_component_types(QS)
+    indices = Int[_query_component_index(query_types, T) for T in types]
+
+    exprs = Expr[]
+
+    if !Unchecked
+        push!(exprs, :(
+            if !is_alive(q._world_state, entity)
+                throw(ArgumentError("can't get components of a dead entity"))
+            end
+        ))
+    end
+
+    push!(exprs, :(@inbounds idx = q._world_state._entities[entity._id]))
+
+    for i in eachindex(types)
+        cols_sym = Symbol("cols", i)
+        val_sym = Symbol("v", i)
+
+        push!(exprs, :($cols_sym = q._storages[$(indices[i])]))
+        if !Unchecked
+            push!(exprs, :(_check_query_component($cols_sym, idx, $(types[i]))))
+        end
+        push!(exprs, :($val_sym = _get_component($cols_sym, idx.table, idx.row)))
+    end
+
+    vals = Symbol[Symbol("v", i) for i in eachindex(types)]
+    push!(exprs, Expr(:return, Expr(:tuple, vals...)))
+
+    return quote
+        @inbounds begin
+            $(Expr(:block, exprs...))
+        end
+    end
+end
+
+@generated function _query_has_components(
+    q::Query{QS},
+    entity::Entity,
+    ::TS,
+    ::Val{Unchecked},
+) where {QS<:Tuple,TS<:Tuple,Unchecked}
+    types = _to_types(TS)
+    _check_no_duplicates(types)
+
+    if length(types) == 0
+        return :(true)
+    end
+
+    query_types = _query_component_types(QS)
+    indices = Int[_query_component_index(query_types, T) for T in types]
+
+    exprs = Expr[]
+
+    if !Unchecked
+        push!(exprs, :(
+            if !is_alive(q._world_state, entity)
+                throw(ArgumentError("can't check components of a dead entity"))
+            end
+        ))
+    end
+
+    push!(exprs, :(@inbounds idx = q._world_state._entities[entity._id]))
+
+    checks = Any[:(_has_query_component(q._storages[$index], idx)) for index in indices]
+    check_expr = foldr((a, b) -> Expr(:&&, a, b), checks)
+    push!(exprs, Expr(:return, check_expr))
+
+    return quote
+        @inbounds begin
+            $(Expr(:block, exprs...))
+        end
+    end
+end
+
+@generated function _query_set_components!(
+    q::Query{QS,CT,OF,RO},
+    entity::Entity,
+    ::Val{TS},
+    values::Tuple,
+    ::Val{Unchecked},
+) where {QS<:Tuple,CT<:Tuple,OF,RO,TS<:Tuple,Unchecked}
+    types = _to_types(fieldtypes(TS))
+    _check_no_duplicates(types)
+
+    if length(types) == 0
+        return :(values)
+    end
+
+    query_types = _query_component_types(QS)
+    indices = Int[_query_component_index(query_types, T) for T in types]
+
+    for i in eachindex(types)
+        if _get_bit(RO, indices[i])
+            throw(ArgumentError(lazy"component $(_format_type(types[i])) is read-only in the query"))
+        end
+    end
+
+    exprs = Expr[]
+
+    if !Unchecked
+        push!(exprs, :(
+            if !is_alive(q._world_state, entity)
+                throw(ArgumentError("can't set components of a dead entity"))
+            end
+        ))
+    end
+
+    push!(exprs, :(@inbounds idx = q._world_state._entities[entity._id]))
+
+    for i in eachindex(types)
+        cols_sym = Symbol("cols", i)
+
+        push!(exprs, :($cols_sym = q._storages[$(indices[i])]))
+        if !Unchecked
+            push!(exprs, :(_check_query_component($cols_sym, idx, $(types[i]))))
+        end
+        push!(exprs, :(_set_component!($cols_sym, idx.table, idx.row, values[$i])))
+    end
+
+    push!(exprs, Expr(:return, :values))
+
+    return quote
+        @inbounds begin
+            $(Expr(:block, exprs...))
+        end
+    end
+end
+
+"""
     close!(q::Query)
 
 Closes the query and unlocks the world.